requirements.txt
# Core dependencies
streamlit>=1.29.0

# PyTorch and vision
torch>=2.2.2
torchvision>=0.17.2
torchaudio>=2.2.2

# Data processing
numpy>=1.26.0
Pillow>=10.0.0

# Image processing
opencv-python>=4.8.0

# Utilities
tqdm>=4.66.0
matplotlib>=3.7.0
pandas>=2.2.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0
seaborn>=0.13.0

# Model management
huggingface_hub>=0.20.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0

# Image processing
opencv-python-headless>=4.8.0
albumentations>=1.3.1

# Development
scikit-learn>=1.4.0
ipykernel>=6.29.0
jupyter>=1.0.0
